Logging refactor

+++ amami/commands/um2nc.py
@@ -17,19 +17,24 @@ import cf_units
 import cftime
 import netCDF4
 import iris
+import iris.util
 import iris.coords
+import iris.exceptions
+import iris.fileformats
 import amami
 import amami.um_utils as umutils
 from amami.um_utils import Stash
+from amami.exceptions import AmamiError, UMError
 from amami.loggers import LOGGER
 from amami.helpers import get_abspath
 
-def get_nc_format(format_arg:str) -> str:
+
+def get_nc_format(format_arg: str) -> str:
     """Convert format numbers to format strings"""
     nc_formats = {
-        1:'NETCDF4',
-        2:'NETCDF4_CLASSIC',
-        3:'NETCDF3_CLASSIC',
+        1: 'NETCDF4',
+        2: 'NETCDF4_CLASSIC',
+        3: 'NETCDF3_CLASSIC',
         4: 'NETCDF3_64BIT',
     }
     try:
@@ -37,7 +42,8 @@ def get_nc_format(format_arg:str) -> str:
     except ValueError:
         return format_arg
 
-def check_ncformat(ncformat,use64bit):
+
+def check_ncformat(ncformat, use64bit):
     """
     Check whether the --64bit option was chosen along with
     the nc format 'NETCDF3_CLASSIC', as they are incompatible.
@@ -49,38 +55,31 @@ def check_ncformat(ncformat,use64bit):
         )
 
 
-def add_global_attrs(
-    infile,
-    fid,
-    nohist
-) -> None:
+def add_global_attrs(infile, fid, nohist) -> None:
     """Add global attributes to converted NetCDF file"""
     if not nohist:
         date = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
         history = f"File {infile} converted with 'amami um2nc' v{amami.__version__} "\
             f"on {date}"
-        fid.update_global_attributes({'history':history})
-    fid.update_global_attributes({'Conventions':'CF-1.6'})
+        fid.update_global_attributes({'history': history})
+    fid.update_global_attributes({'Conventions': 'CF-1.6'})
+
 
 def get_heaviside_uv(cubes):
     """Get heaviside_uv field if UM file has it, otherwise return None"""
     for c in cubes:
         if Stash(c.attributes['STASH']).itemcode == 30301:
             return c
-    return None
+
 
 def get_heaviside_t(cubes):
     """Get heaviside_t field if UM file has it, otherwise return None"""
     for c in cubes:
         if Stash(c.attributes['STASH']).itemcode == 30304:
             return c
-    return None
 
-def apply_mask(
-    cube,
-    heaviside, 
-    hcrit
-):
+
+def apply_mask(cube, heaviside, hcrit):
     """
     Apply heaviside function to cube
     """
@@ -92,7 +91,7 @@ def apply_mask(
     if cube.shape == heaviside.shape:
         # If the shapes match it's simple
         # Temporarily turn off warnings from 0/0
-        with np.errstate(divide='ignore',invalid='ignore'):
+        with np.errstate(divide='ignore', invalid='ignore'):
             cube.data = np.ma.masked_array(
                 cube.data/heaviside.data,
                 heaviside.data <= hcrit
@@ -110,19 +109,19 @@ def apply_mask(
             h_tmp = heaviside.extract(constraint)
             # Double check they're actually the same after extraction
             if not np.all(c_p == h_tmp.coord('pressure').points):
-                LOGGER.error(
-                    'Unexpected mismatch in levels of extracted heaviside function.'
-                )
-            with np.errstate(divide='ignore',invalid='ignore'):
+                raise UMError(
+                    "Unexpected mismatch in levels of extracted heaviside function.")
+
+            with np.errstate(divide='ignore', invalid='ignore'):
                 cube.data = np.ma.masked_array(
                     cube.data/h_tmp.data,
                     h_tmp.data <= hcrit
                 ).astype(np.float32)
         else:
-            LOGGER.error(
-                "Unable to match levels of heaviside function to variable "
-                f"{Stash(cube.attributes['STASH']).long_name}."
-            )
+            long_name = Stash(cube.attributes['STASH']).long_name
+            msg = f"Unable to match levels of heaviside function to variable {long_name}."
+            raise UMError(msg)
+
 
 def apply_mask_to_pressure_level_field(
     cube,
@@ -137,9 +136,7 @@ def apply_mask_to_pressure_level_field(
     """
     itemcode = stash.itemcode
     # Heaviside_uv
-    if ((30201 <= itemcode <= 30288)
-    or
-    (30302 <= itemcode <= 30303)):
+    if (30201 <= itemcode <= 30288) or (30302 <= itemcode <= 30303):
         if heaviside_uv:
             LOGGER.info(
                 f"Masking field '{stash.long_name}' using heaviside_uv field "
@@ -149,11 +146,11 @@ def apply_mask_to_pressure_level_field(
             apply_mask(cube, heaviside_uv, hcrit)
         else:
             LOGGER.warning(
-            "Heaviside_uv field needed for masking pressure level data "
-            f"is not present. The field '{stash.long_name} -- ITEMCODE:{itemcode}' "
-            f"will be skipped.\n"
-            "If you still want convert this field without masking, "
-            "use the '--nomask' option."
+                "Heaviside_uv field needed for masking pressure level data "
+                f"is not present. The field '{stash.long_name} -- ITEMCODE:{itemcode}' "
+                f"will be skipped.\n"
+                "If you still want convert this field without masking, "
+                "use the '--nomask' option."
             )
             return False
     # Heaviside_t
@@ -167,20 +164,17 @@ def apply_mask_to_pressure_level_field(
             apply_mask(cube, heaviside_t, hcrit)
         else:
             LOGGER.warning(
-            "Heaviside_t field needed for masking pressure level data "
-            f"is not present. The field '{stash.long_name} -- ITEMCODE:{itemcode}' "
-            f"will be skipped.\n"
-            "If you still want convert this field without masking, "
-            "use the '--nomask' option."
+                "Heaviside_t field needed for masking pressure level data "
+                f"is not present. The field '{stash.long_name} -- ITEMCODE:{itemcode}' "
+                f"will be skipped.\n"
+                "If you still want convert this field without masking, "
+                "use the '--nomask' option."
             )
             return False
     return True
 
-def name_cube(
-    cube,
-    stash,
-    simple
-):
+
+def name_cube(cube, stash, simple):
     """
     Assign different name properties to cube
     """
@@ -232,6 +226,7 @@ def name_cube(
     if not cube.long_name and stash.long_name:
         cube.long_name = stash.long_name
 
+
 def fix_cell_methods(cube):
     """Remove intervals in cell methods as it's not reliable."""
     # Input is tuple of cell methods
@@ -242,9 +237,11 @@ def fix_cell_methods(cube):
             # Skip the misleading hour intervals
             if i.find('hour') == -1:
                 newi.append(i)
-        n = iris.coords.CellMethod(m.method, m.coord_names, tuple(newi), m.comments)
+        n = iris.coords.CellMethod(
+            m.method, m.coord_names, tuple(newi), m.comments)
         newm.append(n)
     cube.cell_methods = tuple(newm)
+
 
 def fix_latlon_coord(cube, grid_type):
     """Get proper lat/lon coordinate names based on cube grid_type"""
@@ -254,9 +251,9 @@ def fix_latlon_coord(cube, grid_type):
                 coord.guess_bounds()
             # For length 1, assume it's global. guess_bounds doesn't work in this case
             elif coord.name() == 'latitude':
-                coord.bounds = np.array([[-90.,90.]])
+                coord.bounds = np.array([[-90., 90.]])
             elif coord.name() == 'longitude':
-                coord.bounds = np.array([[0.,360.]])
+                coord.bounds = np.array([[0., 360.]])
     try:
         lat = cube.coord('latitude')
         # Force to double for consistency with CMOR
@@ -272,9 +269,10 @@ def fix_latlon_coord(cube, grid_type):
             (lat.points[0] == -90 and grid_type == 'EG')
             or
             (
-            np.allclose(-90.+np.abs(0.5*(lat.points[1]-lat.points[0])), lat.points[0])
-            and
-            grid_type == 'ND'
+                np.allclose(-90.+np.abs(0.5 *
+                            (lat.points[1]-lat.points[0])), lat.points[0])
+                and
+                grid_type == 'ND'
             )
         ):
             lat.var_name = 'lat_v'
@@ -287,20 +285,20 @@ def fix_latlon_coord(cube, grid_type):
             (lon.points[0] == 0 and grid_type == 'EG')
             or
             (
-            np.allclose(np.abs(0.5*(lon.points[1]-lon.points[0])), lon.points[0])
-            and
-            grid_type == 'ND'
+                np.allclose(
+                    np.abs(0.5*(lon.points[1]-lon.points[0])), lon.points[0])
+                and
+                grid_type == 'ND'
             )
         ):
             lon.var_name = 'lon_u'
         else:
             lon.var_name = 'lon'
     except iris.exceptions.CoordinateNotFoundError:
-        LOGGER.error(
-            "File can not be processed. UM files with time series are currently not supported.\n"
-            "Please consider converting using convsh "
-            "(https://ncas-cms.github.io/xconv-doc/html/example1.html)."
-        )
+        msg = ("File cannot be processed. UM files with time series currently unsupported. Consider"
+               " converting with convsh https://ncas-cms.github.io/xconv-doc/html/example1.html")
+        raise UMError(msg)
+
 
 def fix_level_coord(cube, z_rho, z_theta):
     """Rename model_level_number coordinates to better distinguish rho and theta levels"""
@@ -319,6 +317,7 @@ def fix_level_coord(cube, z_rho, z_theta):
         c_height.var_name = 'theta_level_height'
         c_sigma.var_name = 'sigma_theta'
 
+
 def fix_pressure_coord(cube):
     """Fix pressure coords"""
     try:
@@ -327,7 +326,7 @@ def fix_pressure_coord(cube):
         plevs.convert_units('Pa')
         # Round coord points otherwise
         # they're off by 1e-10 which looks odd in ncdump
-        plevs.points = np.round(plevs.points,5)
+        plevs.points = np.round(plevs.points, 5)
         # If needed, flip to get pressure decreasing as in CMIP6 standard
         if plevs.points[0] < plevs.points[-1]:
             cube = iris.util.reverse(cube, 'pressure')
@@ -335,12 +334,14 @@ def fix_pressure_coord(cube):
         pass
     return cube
 
+
 def to32bit_data(cube):
     """Change data to 32 bit"""
     if cube.data.dtype == 'float64':
         cube.data = cube.data.astype(np.float32)
     elif cube.data.dtype == 'int64':
         cube.data = cube.data.astype(np.int32)
+
 
 def set_missing_value(cube):
     """
@@ -352,9 +353,10 @@ def set_missing_value(cube):
         fill_value = 1.e20
     else:
         # Use netCDF defaults
-        key=f"{kind}{cube.data.dtype.itemsize:1d}"
+        key = f"{kind}{cube.data.dtype.itemsize:1d}"
         fill_value = netCDF4.default_fillvals[key]
     cube.attributes['missing_value'] = np.array([fill_value], cube.data.dtype)
+
 
 def convert_proleptic_calendar(cube):
     """
@@ -363,7 +365,8 @@ def convert_proleptic_calendar(cube):
     """
     def _convert_proleptic(time):
         """Convert units from hours to days and shift origin from 1970 to 0001"""
-        newunits = cf_units.Unit("days since 0001-01-01 00:00", calendar='proleptic_gregorian')
+        newunits = cf_units.Unit(
+            "days since 0001-01-01 00:00", calendar='proleptic_gregorian')
         # Need a copy because can't assign to time.points[i]
         tvals = np.array(time.points)
         if time.bounds is not None:
@@ -382,7 +385,7 @@ def convert_proleptic_calendar(cube):
                 date.second
             )
             tvals[i] = newunits.date2num(newdate)
-            if has_bnds: # Fields with instantaneous data don't have bounds
+            if has_bnds:  # Fields with instantaneous data don't have bounds
                 for j in range(2):
                     date = time.units.num2date(tbnds[i][j])
                     newdate = cftime.DatetimeProlepticGregorian(
@@ -417,12 +420,14 @@ def convert_proleptic_calendar(cube):
             new_calendar = 'proleptic_gregorian'
         else:
             new_calendar = time.units.calendar
-        time.units = cf_units.Unit("days since 1970-01-01 00:00", calendar=new_calendar)
+        time.units = cf_units.Unit(
+            "days since 1970-01-01 00:00", calendar=new_calendar)
         time.points = time.points/24.
         if time.bounds is not None:
             time.bounds = time.bounds/24.
     cube.remove_coord('forecast_period')
     cube.remove_coord('forecast_reference_time')
+
 
 def cubewrite(cube, sman, compression):
     """Write cube to file"""
@@ -436,7 +441,7 @@ def cubewrite(cube, sman, compression):
                 tdim = tdim[0]
                 neworder = list(range(cube.ndim))
                 neworder.remove(tdim)
-                neworder.insert(0,tdim)
+                neworder.insert(0, tdim)
                 LOGGER.warning(
                     "Incorrect dimension order for ITEMCODE: "
                     f"{Stash(cube.attributes['STASH']).itemcode}.\n"
@@ -444,7 +449,7 @@ def cubewrite(cube, sman, compression):
                 )
                 cube.transpose(neworder)
         else:
-            cube = iris.util.new_axis(cube,cube.coord('time'))
+            cube = iris.util.new_axis(cube, cube.coord('time'))
         sman.write(
             cube,
             zlib=True,
@@ -460,6 +465,7 @@ def cubewrite(cube, sman, compression):
             fill_value=fill_value
         )
 
+
 def main(args):
     """
     Main function for `um2nc` subcommand
@@ -472,23 +478,24 @@ def main(args):
     outfile = get_abspath(args.outfile, checkdir=True)
     # Get netCDF format
     nc_format = get_nc_format(args.format)
-    check_ncformat(nc_format,args.use64bit)
+    check_ncformat(nc_format, args.use64bit)
     # Use mule to get the model levels to help with dimension naming
     LOGGER.info(f"Reading UM file {infile}")
     ff = umutils.read_fieldsfile(infile)
     try:
         cubes = iris.load(infile)
     except iris.exceptions.CannotAddError:
-        LOGGER.error(
-            "UM file can not be processed. UM files with time series currently not supported.\n"
-            "Please convert using convsh (https://ncas-cms.github.io/xconv-doc/html/example1.html)."
-        )
+        msg = ("UM file can not be processed. UM files with time series currently not supported."
+               "Convert with convsh https://ncas-cms.github.io/xconv-doc/html/example1.html.")
+        raise UMError(msg)
+
     # Get order of fields (from stash codes)
     stash_order = list(dict.fromkeys([f.lbuser4 for f in ff.fields]))
     LOGGER.debug(f"{stash_order=}")
     # Order the cubelist based on input order
-    cubes.sort(key = lambda c:
+    cubes.sort(key=lambda c:
                stash_order.index(c.attributes['STASH'].section*1000 + c.attributes['STASH'].item))
+
     # Get heaviside fields for pressure level masking
     if not args.nomask:
         heaviside_uv = get_heaviside_uv(cubes)
@@ -501,6 +508,7 @@ def main(args):
     z_theta = umutils.get_sealevel_theta(ff)
     # Write output file
     LOGGER.info(f"Writing netCDF file {outfile}")
+
     try:
         with iris.fileformats.netcdf.Saver(outfile, nc_format) as sman:
             # Add global attributes
@@ -523,7 +531,7 @@ def main(args):
                     )
                     continue
                 # Name cube
-                name_cube(c,stash,args.simple)
+                name_cube(c, stash, args.simple)
                 # Remove unreliable intervals in cell methods
                 fix_cell_methods(c)
                 # Properly name lat/lon coordinates
@@ -553,7 +561,9 @@ def main(args):
                     f"Writing field '{c.var_name}' -- ITEMCODE: {itemcode}"
                 )
                 cubewrite(c, sman, args.compression)
-    except Exception as e: #If there is an error, remove the netCDF file created
-        os.remove(outfile)
-        LOGGER.error(e)
-    LOGGER.info("Done!")
+
+    # Catch any errors and remove the output file if it exists
+    except Exception as ex:
+        if os.path.exists(outfile):
+            os.remove(outfile)
+        raise AmamiError(ex)
